shell: make Expand work with "~"
This specific case was broken, as the word included a trailing newline,
thus meaning that we were actually expanding "~\n". The interp package
was then trying to figure out the username "\n".

The solution is to strip the unwanted newline before the interpreter
runs, not after.

Thanks to Andrey Nering for spotting this edge case.