syntax: add support for Bash's {varname} redirects
From 'man bash':

	Each redirection that may be preceded by a file descriptor
	number may instead be preceded by a word of the form {varname}.

As with file descriptor numbers, the parser needs some help from the
lexer to avoid backtracking.

When the parser is in POSIX or mksh mode, the parser will error when
finding these redirections. This technically breaks with POSIX, as we
should parse {varname} as an argument to the command, but that may be
confusing and lead the users into thinking that the Bash feature will
work on other POSIX shells.

We already break with POSIX in a few special edge cases, so this seems
like another one that is worth the trouble. It seems extremely unlikely
that anyone would write {varname}<file and expect {varname} to be an
argument, anyway.

The extra work only happens if we see a literal right before a
redirection operator, so performance is unaffected in normal scenarios.

Fixes #257.