syntax: simplify Printer.Print
We no longer appear to need to set p.line to the position of the first
node. This was initially done to ensure that printing single statements
doesn't include leading newlines, but those tests still pass.

It is also cleaner to avoid that hack, as it is dubious what the start
position of a statement list is, due to comments.