syntax: fix parsing of ((a/=b))
That is a valid program, as the slash breaks the first identifier. We
were parsing it as 'a/ = b', which is wrong.

Found by chance while checking if the printer's minification broke any
programs. While at it, add the test that checks that none of the known
test cases are broken when minified.

Updates #259.