syntax: ensure that "$\\\nà" gets parsed correctly
It is a tricky case, as "$à" is not a valid parameter expansion, and
should be parsed as a literal. On top of that, we have an escaped
newline, so the lexer can't know to parse the "$" as a literal
immediately.

Thankfully, we already handled this case correctly. But not the end
position for the word, whose offset was one byte less than expected. The
cause of the issue was that we weren't always resetting p.w to 1 when
the parser hit an EOF. In most cases that worked out OK as we only
parsed ASCII, but this test case doesn't.