interp: remove the need to call Reset typically
There was never a good reason for this, and it had remained unchanged
purely due to laziness to redesign the API.

Simply drop this restriction without redesigning the rest of the API
just yet. Most users of the interp package should now be able to stop
calling Reset.

While at it, make Runner.Fields return an error, since there might be an
error when setting up the Runner or when expanding the fields
themselves.