syntax: stop the lexer after tokenizing backquotes
The lexer needs to know how many open backquote command substitutions
are open in order to parse backslashes properly. However, the lexer does
not have that context - the parser does.

This was showing up if an opening backquote was followed by another
opening backquote, the second being escaped:

	`\` cmds... \``

The program above worked if there was a space between the first two
characters, since the parser opened the command substitution once the
first backslash had already been skipped.

Delay reading the rune right after every backquote character and do that
in the parser, so that the parser can first update the number of open
backquotes.

This adds yet another direct line between the lexer and parser,
unfortunately. There is not much we can do about this, though, as that's
how the shell language works.

Fixes #278.