interp: get rid of Runner.Stmt
This extra piece of API was introduced exclusively for cmd/gosh, since
it needed to know if an ExitStatus error came from an actual exit call
or not. This is because exit codes coming from most commands should not
exit the shell, but those coming from the "exit" builtin should.

Instead, split the ExitStatus error type into two, adding
ShellExitStatus. The latter means exit statuses that actually stop the
shell.

While at it, update the Runner.Run godoc.