interp: introduce a constructor with option funcs
This way, we can have the constructor properly set up a Runner, without
that work falling to Reset. Then, Reset becomes much simpler and no
longer needs to return an error.

Runner.FromArgs also becomes an option function, as it fits nicely into
the signature. It no longer returns the remaining arguments, but its
only use was to simply set those arguments as the shell's parameters.

Finally, adapt all of the test code, as well as other packages like
cmd/gosh and shell.