syntax: support comments on elses and cases
They already parsed and printed without errors, but they were sometimes
misplaced. For example, in the piece of code below, we'd group the two
comments together, and indent both of them as if they were part of the
"if" body:

	if foo; then
		# inside the body
	# document the else
	else
	fi

This now works better. Any trailing comments that are aligned with the
opening token - else, elif, case - now get attached with it and aligned
with it.

Fixes #279.