syntax: improve how we stop parsing regex words
We can't exit the test regular expression state after we've parsed the
entire word. That can lead to mis-parsing characters that follow the
regular expression. In the added example, a || test operator was parsed
as a literal instead of a token.

We need to teach the lexer more about test regular expressions, so that
it knows to get rid of the state before reaching the || operator.

This fix isn't perfect; see the TODO. However, the new code isn't as
broken as the old one, and some common inputs like the added test now
pass.

Fixes #287.