interp: use globbing more conservatively
Not all strings that require wildcard pattern quoting can expand to
arbitrary paths on disk. For example, "echo \\" needs quoting, but must
always output "\\".

This was an issue especially on Windows, as the above command was
globbed and matched the current directory, meaning that "." was printed.

To properly support this, we need a new piece of API in the syntax
package. What we need cannot be done via neither of the two existing
wildcard pattern APIs, which is why we had a bug in the first place.

Fixes #291.