CI: remove test -race from AppVeyor
It's now failing because gcc isn't installed. We previously used a Go
installation downloaded from Google, and now we use the pre-installed
version from AppVeyor. Perhaps it's set up wrong, in a way that the C
compiler is misconfigured.

All in all, most code is OS-independent, so we don't really need to run
the race detector on Windows. We already run it on Linux.