expand: new package, adopting syntax.ExpandBraces
Code to statically expand nodes and strings should not live in the
syntax package. It clutters the package's API, and also has little to do
with parsing shell syntax.

We can't move the code itself just yet, as we can't remove
syntax.ExpandBraces and that would result in a cyclic dependency.

While at it, update the experimental package godoc notices.