cmd/gosh: add first interactive shell tests
The code to implement an interactive shell isn't trivial, so we need
tests to ensure all the edge cases are properly covered and not broken
over time.

This required some refactor in the internal test types, as well as less
hard-coding of the system's input and output writers in the gosh
package.

For #295.