interp: start decoupling expansion from interpreting
Most of the logic to expand syntax nodes into words and strings cares
little about interpreting. All it needs is some environment or context.

Tightly coupling the two from the start was a mistake. This means that
it's very hard to perform expansions without setting up a runner and
potentially running arbitrary code.

In mvdan.cc/sh/shell we introduced a "pure runner" to try to mitigate
this, but that's still far from perfect. It's more complex, pulls in
more code, and still has to deal with arbitrary shell code.

Decoupling the two halves of the interp package will be hard, but it's
possisble. This is the first step; separating the two structs, leaving
some TODOs and rough APIs.

Follow-up commits will continue this work.