interp: simplify handling of IFS
Joining with the first character of IFS via "*" isn't common, so we
don't need to optimize for it. Similarly, ifsRune doesn't need to be a
function literal; a method will do.

With the two changes above, it's simple to remove necessary input fields
in ExpandContext. It now simply queries IFS via Environ every time one
of its entry point functions is run, which is a cheap operation.

While at it, moved ifsFields to expand.go, since it belongs there.