expand: move shell expansion code over from interp
Since we had done a few incremental changes already, only a few pieces
of work were left to do for this to happen.

Leave some functions like match and flattenAssign back into interp, as
they aren't useful enough to be exposed as APIs in the expand package.

Make the Subshell expand field take an io.Writer instead of returning a
string, as that means we can reuse the internal buffer and avoid
allocations.

Finally, some more APIs had to be exposed, such as Literal and Pattern.
The new names are more adequate than their old unexported counterparts.

The API is far from done; we'll likely rename most exported names, and
godocs are missing.