expand: minor tweaks to the exported APIs
All expansion that happens to a list of words when executing commands
should take ...*Word, and not a single word, for consistency and
simplicity. Most uses will use a slice, but we still want to support
one-off uses.

In this case, modify Braces to be consistent with ExpandFields.

While at it, make ExpandFormat return the output string first, as it's
the main return value. The integer is just extra information about it;
the number of arguments used.