expand: refactor Environ, add more constructors
Remove Sub from the main interface; it's for a really niche use case and
it doesn't seem particularly useful at the moment. We should get back to
that once we support global and local variables properly.

Now that Environ implementations don't have to worry about subshells and
nested scopes, we can expose the func and list Environ implementations
in the expand package better. While at it, make ListEnviron remove
duplicate and invalid entries from the input list.

Keep interp.mapEnviron as overlayEnviron, as we still need to use
overlays on top of the global environment for a few use cases.