interp: remove TODO about the system's environment
We're not going to do this after all. Each call to os.Environ makes a
full copy, so the Each method would become more expensive.

If anyone needs live changes to the environment to affect the shell
library, they can use FuncEnviron(os.Getenv), or implement a mix of
os.LookupEnv and os.Environ in a dozen lines of code.