syntax: keep <<- indentation in the syntax tree
Otherwise, it's impossible to know what the existing indentation was
like. Leave it up to the user (such as the printer or the interpreter)
to remove them or replace them, if needed.

This does make the interpreter a bit more complex, but we can make the
parser simpler and the syntax tree more precise.

For #309.