shell: make source funcs take a context
Instead of enforcing a timeout of exactly two seconds. A user may want a
different timeout or a cancel function, and enabling that doesn't
complicate the signature much.

If anything, it's expected that any function that could potentially take
a long time should take a context parameter.

While at it, add a test for cancelling a SourceFile call.