expand: report env vars as exported
Generally, all variables set up for expand.Environ should be exported,
so that they affect executed programs. Do this for ListEnviron and
FuncEnviron.

Adapt the interp package to apply those rules, and do the same with
inline variables in simple commands.

Finally, add a test where we override an exported environment variable,
which was failing before because we didn't see said environment
variable as exported.