syntax: add Word.Lit helper method
This is quite a common operation, and it's not as trivial as it may
initially seem. For example, the "foo$" word contains two literals, as
the parser can't directly see that the dollar sign doesn't start a
parameter expansion.

This allows us to simplify a bit of code in the expand package, which
was rolling its own solution.