expand: remove WriteEnviron.Delete
It overlapped with Set a bit, since Set(name, Variable{}) didn't make
sense before.

It also wasn't used by the expand package at all, so it didn't really
belong in that interface.

interp.Runner no longer makes direct modifications to the provided
Environ, so it's no longer needed there either.

We might change the semantics of Set slightly in the future, such as how
to export an existing variable. For now, the current interface and
semantics will do.