shell: remove bad sandboxing from SourceNode
Limiting what programs and files can be run may be useful in avoiding
some malicious code, but it is never going to be perfect. For example,
we allowed running the "sed" program, which has an option to modify a
file on disk.

Running arbitrary native shell code can also be problematic. For
example, one can spin the CPU until the context deadline is reached, or
one can try to bring the host down with a fork-bomb.

These are just simple examples; even if they were covered, there could
be countless others. There's simply no point in trying to have a pure
sourcing API.

If one wants to implement simple protection against running arbitrary
programs or opening arbitrary files, that can still be done like the old
code demonstrated. However, we shouldn't expose that as a "safe" or
"pure" API.