interp: stabilize API
We're going to freeze v2 and start work on v3 (see #266), so it makes
sense to stabilize the interp package now that we're reasonably happy
with its API.

The API is not perfect, but it's definitely better than earlier v2
versions. And plenty of packages already depend on the existing API,
anyway.

Fixes #283.