Revert "syntax: handle escaped newlines in a single place"
Handling escaped newlines in the Parser.rune method was a good idea on
paper, but in practice it was buggy. For example, in the presence of two
backslashes followed by a newline, the new code would get confused and
think that the newline was escaped.

A follow-up commit will add some regression tests that had been broken
in that commit.

This reverts commit 71b31cb0ccf556501f5e69fcde17a8dbe479b491.