expand,shell: refactor expansion of words
The expand package needs more API than just Fields, Literal, and Pattern
to expand words. One API was missing; expanding a word as if it were
within double quotes, which is what happens in a here-document, and what
shell.Expand needs.

As a result, interp's here-documents no longer incorrectly expand
tildes - and the same can be said about shell.Expand.

While at it, improve and clarify the docs.

Thanks to @rogpeppe for his help in figuring out the right API for the
shell package.