CI: move Windows testing to Travis
This lets us drop AppVeyor, which means one less YAML file to maintain,
and one less service to know how to use.

Windows on Travis is still in early release and slow, but it mostly
works. Compared to AppVeyor, it's not particularly slow or hard to get
working; if anything, it's simpler thanks to the better Go support.

The only two issues encountered were the CRLF line endings in the
canonical.sh test, and the lack of the "env" program. Both have been
worked around for now.

Fixes #299.