expand: avoid var lookup loops in Arithm
We tried to expand every literal as a variable. This kept the code busy
forever if it encountered a loop, such as 'set -- 1' or 'a=b b=a'. Add a
recursion limit, and add some test cases.

While at it, add a shortcut - if a value is not a valid name, there's no
point in trying to expand it as a variable.

Fixes #317.