interp: fix unsetting local variables
This action should not modify global variables.

Another wrinkle also needs fixing; in Bash, unsetting a local variable
does not completely get rid of the local definition. For example,
assigning to the unset name does still not modify a previously declared
global variable.

To fix that last wrinkle, we must relax Variable.IsSet to allow
Variable.Local to be true but have the variable still be unset.
Otherwise, there's no way for Environ.Get to return a local variable
that is unset, something which is necessary for the edge case behavior
described above.

And added a few test cases for these wrinkles.

Fixes #318.