expand: remove Config.OnError
This started as a way to fit expand.Config into interp.Runner in an
easier way, but it's lazy. For the user, reporting errors via return
parameters is much more Go-like and simpler to use.

This also means that it's simpler to reason about the control flow. It's
now easy to tell if an error will cause an expand operation to exit
immediately.

Add a few method wrappers in interp.Runner, to keep the code simple and
readable. The expand code does get a bit more verbose, but that's a good
tradeoff to make.