all: transition to a Go module
We can't safely convert mvdan.cc/sh v2 to a Go module, as that would
require anyone doing minor or bugfix updates to fix their import paths.
Not only is this a breaking change, but would also be impossible on
earlier Go versions like 1.10 or 1.10.2.

We want the project to be a Go module, since that will provide many
advantages. So take the opportunity and start work on v3. We'll release
v3.0.0-alpha0 right away, so that users can depend on that semantic
version.

For now, the transition to a Go module will stay in the master.v3
branch, to not break existing users - especially those using 'go get' to
install binaries. We will likely replace master with master.v3 once Go
1.12 has been released, since we'll drop Go 1.10 support then.

Updates #266.