all: drop Go 1.10, have CI build in module mode
go.mod already requires GO 1.11, so this makes it consistent. Using the
module build on CI simplifies the setup there, and ensures that CI uses
the same versions of all dependencies that we do.

The 'go get -d -t ./...' should no longer be necessary, as 'go install'
and 'go test' will download dependencies as needed.