syntax: account for literals in Parser.Incomplete
In some edge cases, such as the one in the added test case,
Parser.Incomplete may incorrectly return false. This is because if we
haven't finished lexing the first literal of a statement, we have yet to
start parsing the statement.

Fix this by making Parser.Incomplete check that we're not currently
lexing any literal.

Fixes #320.