syntax: add BraceExp as an exported WordPart
This means we can make use of other pieces of code that rely on Word and
WordPart, such as Word.Lit. We can also get rid of the extra boilerplate
and interface{} usage.

But most importantly, this is the first step towards allowing syntax
users to inspect brace expansions without expanding them. That will come
in a follow-up commit.