interp: add support for the $(<file) shortcut
Bash supports this so that one can get the contents of a file without
executing a program like cat. However, supporting it requires a special
case - a statement that simply uses a file as standard input shouldn't
do anything besides just opening the file.

Add that special case, and a few tests.

Fixes #323.