README: link to official docker image
Keep the third party docker image links, for now. Also say "packages
available on" instead of "available for", since some of the listed names
are package managers and not operating systems or distros.

cmd/shfmt/Dockerfile was added after the v2.6.1 tag was done, so this
one time I've built and pushed the image manually. Future builds should
happen automatically on Docker Hub.

The tag names are like "v2.6.1" and not "2.6.1", since that's the same
version name used in the git tags and by 'shfmt -version'.

Fixes #68.