expand: uppercase names in ListEnviron on Windows
We were doing this in the interp package before the code was extracted
into the expand package. Getting rid of that was a mistake, and led to
subtle bugs on plain Windows.

To prevent this from happening again, add a regression unit test, as
well as a full interpreter test via os.Setenv in the interp package.

Fixes #325.