syntax: better unicode support in the parser
Before, we would assume that all non-ASCII runes would need at least
four bytes to be parsed properly. This was fine while the parser wasn't
meant to be used interactively, as forcing a read too early wasn't a big
deal.

However, now that we have Parser.Interactive, forcing an extra read can
mean that the user hits enter and is prompted for extra bytes even if
they typed a full command.

Instead of working on assumptions, use utf8.FullRune to see if we have
the start of an utf8 rune, but we're missing bytes. Only in that case
should we read more bytes.

Fixes #327.