syntax: don't skip Hdoc in Redirect.End
If a redirect has a heredoc body, its end should be the heredoc body's
end. This was the source of the mysterious bug in walk_test.go and the
attached TODO.

Two minor changes in the printer and walker were necessary, to take into
account the slight change in semantics of Redirect.End.

Node.End has always been documented to point at the first character
"after the node", so the previous behavior was broken.