syntax: ValidName should return false on ""
An empty string isn't a valid name. For example, one cannot do any of
the following:

	declare =bar
	echo ${}

This wasn't a big problem in the parser, as it already knew whether or
not it had a non-empty literal. But other users of ValidName, like
interp, would get bugs because of the missing empty string check.