all: refactor how Docker images are built
Now the Dockerfile is version agnostic, so it can be used to build a
master version as well as any tagged version.

It also uses COPY instead of a git clone, so it's much faster and no
longer requires an internet connection.

These changes mean that we can no longer use Docker Hub's automated
builds. Which is fine, because they were painfully slow anyway.

While at it, add an ENTRYPOINT as suggested in #68.