syntax: make StmtList a Node
It's not particularly useful yet, but it will be useful very soon to
clean up IfClause.

We also did have Pos and End methods for this type, though we hadn't
exported them. Removing the type or the methods would be a net loss of
simplicity in the syntax tree type definitions, so properly expose the
type and methods.

It's the first node whose Pos and End methods don't take pointer
receivers. That's fine, as it is often used as an anonymous field.