expand: preserve glob prefixes such as ./
For example, "./foo/*" should match "./foo/bar", not "foo/bar". This can
matter for some programs, such as "go build ./foo/*", since the leading
dot directory indicates a path on disk as opposed to a URL to be fetched
via the network.

We didn't support this well because we were using filepath.Join
everywhere, which cleans paths, thus removing dot directory components.

To fix this, avoid joining paths whenever possible, and use custom logic
when joining to avoid cleaning the result.

Thanks to Andrey Nering and Marc Trudel for finding this bug in the
go-task/task project.

Fixes #336.