expand: re-add slash conversion to globbing
When we removed all uses of filepath.Join and thus filepath.Clean, we
forgot that Clean also performs FromSlash. We need that piece, because
we want globbing with forward slashes to work on Windows too.

This is useful so that a unix-like script with globbing also works on
Windows. There's no reason not to do that, since slashes are forbidden
from Windows file and directory names, anyway.

The regression was introduced in the previous commit, and is now fixed.