expand: resolve the two remaining TODOs
The main point of hasGlob in older v2 versions was to avoid doing
globbing work for many more directories than needed, since we'd perform
globbing on filepath.Join($PWD, field).

However, newer versions glob only on the field relative to PWD, so we
already skip that extra work without the need for this piece of code.

As for the binary search, it can be implemented in just a few lines
thanks to sort.SearchStrings.