syntax: pipes are left-associative too
That is, the program:

	foo | bar | baz

Should parse with recursive pipes at the left, like:

	       |
	    |     baz
	foo   bar

We already did this for && and ||, but not pipes. This makes the syntax
parser more consistent and correct.

We noticed now because we added an optimization where a shell's pipe
would reuse copy buffers. With right associativity, the original shell
performs all pipes, so the same buffer was used by multiple copies at
the same time, resulting in bugs.

In other words, a shell pipe forks the left side into a subshell, so
nested pipes should go to the subshell.