expand: work around incorrect globbing error
v3, currently in the master.v3 branch, doesn't have this bug as the
globbing code that calls ReadDir was refactored. This older code
performs two ReadDir calls, and one of them is run even if the current
potential glob path is not a dir.

This can be a problem if one performs a glob like "*/foo" and the file
"bar" exists; we should skip it, instead of erroring because it's not a
directory.

Fix that in a simple way in v2, and add a regression test.

Fixes #347.