expand,interp: adapt to Bash 5.0 changes
The order in which associative array keys and values are printed has
changed. Before it was sorting by key; now it appears to sort in an
arbitrary way, such as via the hash table.

'man bash' still doesn't document what the order is, so let's keep the
simple order and have the tests sort manually.

'${!a}' now errors if the variable is empty, so do that too.

Updates #346.