shell: deflake TestSourceFileContext
It was cancelling the context after firing the goroutine. If the host
machine is busy enough, or if the user is unlucky enough, it could be
that the fired goroutine finishes before the parent has cancelled the
context. Thus leading to a nil dereference panic as we'd call .Error on
a nil error.

To fix this, remove the goroutine race condition, and stop assuming that
the error will be non-nil.