syntax: rethink IfClause for v3
Following a TODO from a while ago, we can remove three fields from
IfClause entirely. The node becomes a bit easier to work with, mainly
because of the reduction of redundant fields.

The most important change is that Else is now an *IfClause, representing
either an "elif" or an "else" if non-nil. The other notable changes is
that FiPos is always valid, and ThenPos is no longer valid in the "else"
case.

Finally, stop the printer from removing empty else branches. It's not
its place to remove pieces of nodes. We might teach syntax.Simplify to
do this in the future.

While at it, add a v3 TODO to update Simplify's signature.