_js: bump to v0.4.0
Versions 0.3.2 and 0.3.3 were published in the master branch, to follow
the latest 2.6.x changes in the mvdan.cc/sh module itself.

The master.v3 branch has had quite a bit of work done that isn't in the
old v2 versions, so let's start updating NPM from this branch.

A few tweaks were necessary to get the module-aware gopherjs building
properly. Otherwise, everything works like it used to in the master
branch.