interp: support lowercase volume names in $PATH
On Windows, "C:" is a valid volume name, just like "c:". We were
incorrectly assuming that the letter had to be uppercase.

Add a simple regression test for Windows, which sets up a PATH with a
script in a directory, then tries to execute it via such a PATH with a
lowercase volume name.

Fixes #355.