cmd/gosh: don't print exit status errors
If the final interpreter error is an exit status, we should use that for
gosh's own exit status and not print the unnecessary error line. This is
the same that Bash and other shells do:

	$ gosh -c 'exit 2'; echo $?
	exit status 2
	1
	$ bash -c 'exit 2'; echo $?
	2

The interactive mode was handling this correctly already. Clean up the
code to be consistent there too.