expand: tilde expansions shouldn't expand further
To quote the latest version of the POSIX Shell spec:

	The pathname resulting from tilde expansion shall be treated as if
	quoted to prevent it being altered by field splitting and pathname
	expansion.

That is, when HOME='/*', the command "echo ~" should just print "/*" and
not do any globbing at all.

This was usually not a problem on Unix-like systems, but Windows uses
backslashes as path separators. Thus, HOME='C:\foo\bar' is a very normal
occurrence.

However, since we treated the result of tilde expansion as unquoted,
"echo ~" would result in the output "C:foobar", since we'd eat the
backslashes when taking care of character escapes.

Fix that, and add a few tests for these edge cases.

Fixes #356.