syntax: add ParseError.Incomplete
During read calls, it was already possible to check if more bytes were
required to finish parsing without errors, via the Parser.Incomplete
method.

However, once an error was returned, that information was not easily
available. Add a field to the error type, and add an IsIncomplete
function to quickly check errors for that property.