_js: add a nested Go module
This is useful so we can pin exact GopherJS versions as requirements,
without adding them to the main module's graph. There are users of the
sh module out there who have no interest in the javascript library, so
they shouldn't be forced to ever see or download GopherJS's repository
or go.mod file.

Also replace the sh/v3 module with ../, so that the _js module is always
using the sh/v3 version from the very same git tree. This also means
there will be fewer network roundtrips, and the _js module will be up to
date by default.

We needed to call the nested module shjs instead of mvdan.cc/sh/v3/_js,
as that caused ambiguities for cmd/go. We actually don't care what the
module path of the JS module is, as it's never going to be importable
directly.