_js: add a non-blocking wrapper of Interactive
JS only has a single thread, meaning that if the interactive parser
blocks on a call, the only way to run any other code in the meantime is
to run it as part of the "read" call.

This is understandably clunky, as the API was designed for concurrent
languages. However, with about twenty lines of code we can offer an API
that never blocks to JS, which is more usable for real programs.

It does re-parse chunks of incomplete statements, but those shouldn't
get very large. Adding a non-blocking Interactive variant to the
original parser would be too much of a refactor, since it's a recursive
descent parser. It can't be stopped and resumed, as the parse state is
in the call stack.