_js: expose IsIncomplete, don't stringify on throw
For example, ParseError contains more information than just the message,
and we want to have that available.

Errors from the JS layer shouldn't be the Go-style error interface. JS
already has an Error type, whith a message string and its own stack
trace. Use that, which is as simple as calling panic in Go.