syntax: get rid of StmtList
Its main purpose was to easily share two fields and a few methods across
the entire syntax package, meaning we wouldn't need to have the same
pairs of fields in a dozen structs.

However, this leaks into the API and makes it ugly. Composite literals
get more complex, and we add the complexity of promoted fields to most
of the nodes.

This is also a problem when exposing the syntax package to other
languages. For example, JavaScript has no promoted fields, so all
accesses to statement lists had to be explicit and verbose.

Instead, get rid of the indirection. The internal syntax code gets
slightly more verbose, but also simpler. And most importantly,
practically all users of the syntax package get their code slightly
simplified.