syntax: fix reserved words immediately followed by redirects
This was broken due to a small oversight; in the added test case below,
"fi>/dev/null" would parse similarly to "2>/dev/null", as if "fi" was a
file descriptor number.

This mistake in the lexer confused the parser, which then proceeded to
complain about "fi" being missing or displaced.

Fixes #369.