interp: record runner state at the first Reset call
This is perhaps a bit confusing, but it's a good middle ground. This
lets us support Reset while not undoing changes that the user has done
outside of New, which is common.

The downside is that making any changes after the first call to Run or
Reset are undone by further calls to Reset. However, that's fine; Reset
is meant to return a runner to its initial state, not any other previous
or arbitrary state.