syntax: support coprocs with non-literal names
From the Bash 5.0 release notes:

	h. The name argument to the coproc reserved word now undergoes
	word expansion, so unique coprocs can be created in loops.

In other words, the name used to just be a name literal, and now can be
any word which expands to a valid name.

On Bash 4.4:

	$ var=name
	$ coproc $var { echo foo; }
	bash4: `$var': not a valid identifier

And on 5.0, we don't get an error:

	$ var=name
	$ coproc $var { echo foo; }

To adapt to this, simply allow any word in that node field, and not just
literals. The interp package doesn't yet support this syntax node, so we
don't have anything else to do.

Fixes #346.