syntax: parse [[ X = Y ]] without changes
The parser used to parse the operator as if the input were [[ X == Y ]],
since that form is more consistent and there's no reason to use '='
instead.

However, this was problematic. It was hard to syntax highlight the
original form, as it was impossible to know what token was used in the
input without looking at the input bytes directly.

Instead, apply this transformation as part of syntax.Simplify. It's now
possible to parse and print the original test expression without
changing the token.

Updates #319.