syntax: don't parse options in DeclClause
It's impossible to do statically. For example, imagine:

	opts="-a -b"
	declare $opts name=value

We already support parsing declare arguments at run-time, so do the same
for options. This unfortunately makes the node less easy to analyze
statically, but that's just how Bash is designed.

On the plus side, the new design is simpler. Now DeclClause is almost a
regular builtin via CallExpr, except that it can also contain arrays.

Fixes #332.