syntax: make SplitBraces work in-place
This makes it consistent with Simplify. Modifying nodes in-place is
better than generating copies, because most of the time the user doesn't
need both in memory at once.

In the rare cases that one does have that need, it's always possible to
print and re-parse the node, as a simple way to do a deep copy.

On the other hand, if the API always makes a full copy by design,
there's no way to work around that extra garbage generation.