interp: remove stat module TODO
I can't think of a good use for it right now. We could replace
OpenModule with an interface that includes some combination of other
methods like Stat, Lstat, Readlink, and Samefile. However, that would
complicate the module.

One nice property about the current OpenModule is that, while being
simple, it allows controlling write access to the system, along with
ExecModule, since OpenFile is the only call that can modify the
filesystem.

Let's leave it as it is for v3 for now. If there's a real need for a
more powerful module in the future, we can always consider it
separately.