syntax: make DblQuoted have Left and Right pos fields
This wasn't strictly necessary in the past, but it now is to properly
handle leading and trailing escaped newlines within double quotes.

This also makes the struct more consistent with all the other nodes that
have opening and closing tokens. Finally, it makes DblQuoted.End much
simpler, which is always nice.