add continuous fuzzing via fuzzit
This mainly means adding the code from the old 'fuzz' branch, adding a
go.mod to track dependencies such as go-fuzz and the fuzzit tool, and
then making CI run the appropriate commands for regression testing and
continuous fuzzing.

Finally, add the badge to show the current status.

We've used a separate module to not pollute the main go.mod and go.sum
files with all the fuzzing-specific deps. This also means that it's
simpler to construct a GOPATH for go-fuzz, as we can simply use 'go mod
vendor' and rename the folder.

While at it, start using GOPROXY on CI to speed up the builds, and
remove GO 1.11, since it doesn't support GOPROXY - and since 1.13 is
about to be released.