interp: fix more incorrect set -e edge cases
For example, we would incorrectly exit the shell if an if condition
failed. Quoting the Single UNIX Specification:

	When this option is on, if a simple command fails for any of the
	reasons listed in Consequences of Shell Errors or returns an
	exit status value >0, and is not part of the compound list
	following a while, until, or if keyword, and is not a part of an
	AND or OR list, and is not a pipeline preceded by the ! reserved
	word, then the shell shall immediately exit.

Would be better if we could quote the POSIX Shell spec, but I wasn't
able to find a quote with a bit of digging. At least bash and dash seem
to agree with the test cases.

Fixes #404.